refactor: tk-encode is the v1 crate; v0 moves to tk-convert
`config` leaves the default feature set. The default `tk-encode` is the v1 crate:
it reads a `.tok`, encodes, and has no serde in its dependency tree at all. v0 —
the legacy `tokenizer.json` reader, the wrapper enums, `Tokenizer` — is what
`config` turns on, and `tk-convert` is the only thing that turns it on.

`tokenizer/tok.rs` keeps only the read half, which is how a v1 build constructs a
pipeline. The writer moves to `tk-convert`, where the conversion belongs: it is
now the only crate that names `Tokenizer` or a wrapper enum.

  cargo tree -p tk-encode -i serde  ->  only via criterion (dev-dependency)

50/50 byte-exact, 332 tests pass with and without `config`.
